move rewriteFileJob to the jobs dir
